requirements.txt
boto3==1.12.8
discord.py==1.3.2
watchtower==0.7.3
requests